The Cleanup Patch
Bumped the version to 2.5.1. Added the cleanup thread that enforces the
DMAX attribute on HOSTS and METRICS. This means that hosts and metrics can
be timed out of the XML tree (and hashes) as specified. The definition is:

'A metric will be deleted DMAX seconds after it is received, and DMAX=0 means
eternal life.'

Currently hosts and builtin metrics will NEVER be deleted, however once
this code is hammered (tested hard) this may change. Gmetric, however, now has the
ability to specify --dmax and it will be enforced.

Other various code-language cleanups have also been incorporated by this patch.