The Subtree Gmetad
This version of gmetad constructs a full DOM tree of the XML, similarly
to gmond. With this structure, it enables an interactive mode where you
can select XML subtrees and summary views. This will help the speed of
the webfrontend's meta, cluster, and host views immensely, among other
things.

I have added a backwards-compatibility mode that will affect version 2.5.0
behavior (no grids). The polling interval bug in conf.c is also fixed and
tested.

This gmetad will still deliver the full tree on port 8651 (by default), and
listens for interactive requests on port 8652.