build: add --with-python support to configure
refactors the logic used to detect the python interpreter so that it works
in platforms that don't have it installed with a /usr prefix.

it defines a --with-python configure flag to help configure locate python
and implicitly enable python support (overrides --disable-python) by
indicating the location of the python interpreter or the prefix where it is
installed.

  svn merge -c1300 trunk
  svn merge -c1303 trunk

tested in DragonFlyBSD, FreeBSD, NetBSD, OpenBSD, Solaris, Cygwin and Gentoo
Linux by me and SuSE Linux by Brad

Reviewed-by: Brad Nicholes <BNICHOLES@novell.com>