web: report correctly uptime in shownode when host is down
complete the fix for BUG36 committed in r1244 to display correctly the time
since the last received heartbeat for a node that is down.

The template for the show_node page has been changed to display a message to
report if the cluster is down (as that will freeze the counter) and to display
the counter using an uptime format if older than 1 hour (as used elsewhere) so
that the observed downtime from a node could be quickly identified.

  svn merge -c1248 trunk
  svn merge -c1249 trunk
  svn merge -c1318 trunk

based on a patch by Jason A. Smith

Reviewed-by: Brad Nicholes <BNICHOLES@novell.com>