gmond: SOURCE for metrics is meaningless but kept for compatibility
after the XDR refactoring there isn't a difference between core metrics 
and user metrics so the SOURCE element for METRIC/METRICS has no more 
meaning and is only needed to avoid breaking backward compatibility at 
the XML layer.

SOURCE="gmetric" is not valid anymore and frontend code that relied on 
it for special handling for user generated metrics is invalid as well.

  svn merge -c 1324 trunk
  svn merge -c 1325 trunk
  svn merge -c 1327 trunk

Reviewed-by: Brad Nicholes <BNICHOLES@novell.com>