web: report local cluster time in node view, remove adhoc message
as reported by Jason A. Smith, the current logic reports a cluster down
even when it is only 1 node.

since the time for the last hearbeat is calculated relatively to the
localtime for the cluster, adding the local cluster time makes more sense
even when the node is up, and will allow the users to understand better
what the time reported means in case of downtime and therefore there is
no need for a message trying to explain that.