web: report correctly uptime in shownode when host is down
as reported by Jason A. Smith where the solution implemented in 3.1 for
BUG36 was mistakenly reporting all the cluster down when only 1 node was
affected.

instead of trying to guess if the cluster is also down to warn the user
that the downtime reported is not accurate because we have no local time
to calculate it from, also show the cluster local time so that the time
reported can be interpreted correctly.

  svn merge -c 1568 trunk
  svn merge -c 1577 trunk

Reviewed-by: Jesse Becker <hawson@gmail.com>