web: propagate use of /dev/null to all rrdtool graph calls for calculations
avoids a PHP notice at least in platforms that have /dev/null by instructing
rrdtool explicitally to use that as the destination for its image generation

reported by Jesse Becker to generating "Undefined offset" notices at least
when used in meta context and fixed in r1754.

it will still fail in windows as there is no /dev/null but the failure will
be similar to the current setup where an invalid file was used instead.