gmond: reformat memory metrics to match pre 3.1 style
with the 3.1 release the memory metrics were moved to use a float instead of
an unsigned but the format for the number used was inadvertly changed to
include 3 decimal digits which are never practically used.

revert to showing no decimal numbers as it was done when the memory metrics
were instead implemented using a 32bit integer.

  svn merge -c 1737 trunk

Reviewed-by: Jesse Becker <hawson@gmail.com>