gmond: add boolean option to 'allow_extra_data' reporting
Allow gmond to be configured not to report the EXTRA_DATA tags to save
bandwith and processing time when they are not being used as proposed
in BUG199.

  svn merge -c 1701 trunk

Suboptimal fix as it only prevents reporting the extra data information
even if gmond is still internally generating it and returns an empty
METRIC tag instead of fully reverting to pre-3.1 XML format.

Reviewed-by: Brad Nicholes <BNICHOLES@novell.com>
Signed-off-by: Carlo Marcelo Arenas Belon <carenas@sajinet.com.pe>
Signed-off-by: Doug Nordwall <doug@pnl.gov>