web: reorder paramater list for graph generation for consistency
if graphargs contains the size of the graph to generate, the location
it is in the URL is significant, as for zoom to work, the size needs
to be overriden with "z=large".

move graphargs to the beginning of the generated URL as it is already
done for the metric graphs in host view.

Signed-off-by: Jason A. Smith <smithj4@bnl.gov>