gmetad: CVE-2009-0241: buffer overflow in interactive port
dynamically allocate the buffer used in process_path to avoid a stack
based buffer overflow when processing path names bigger than 256 bytes.

also avoid an "off-by-one" overflow which could result in the leak of
filehandles when processing requests of up to 2048 bytes and avoid
returning the full tree if the request is not valid in preparation
for future work that will refactor the protocol used by the interactive
port.

  svn merge -c 1946 trunk;
  svn merge -c 1948 trunk;
  svn merge -c 1950 trunk;
  svn merge -c 1953 trunk;
  svn merge -c 1955 trunk;
  svn merge -c 1956 trunk;

Signed-off-by: Spike Spiegel <fsmlab@gmail.com>
Signed-off-by: Kostas Georgiou <k.georgiou@imperial.ac.uk>
Signed-off-by: Carlo Marcelo Arenas Belon <carenas@sajinet.com.pe>
Reviewed-by: Brad Nicholes <BNICHOLES@novell.com>