gmond: ensure Python.h gets included before any standard headers
indicated as a requirement by Python/C API as shown in :
	        
  http://docs.python.org/3.0/c-api/intro.html#include-files

reported by Sergei Volkov in BUG228 to generate a broken modpython
for Solaris.

committed to trunk in r1992 and confirmed to solve the reported
problem in Solaris 10 update 3 and 6 linking against python 2.3
and 2.4; also cleaning compilation warnings in other platforms.