libmetrics: linux: avoid stack overflows when parsing very long mounts
reported to overflow regularly when using the NFS implementation in
RHEL5 that creates temporary mountpoints entries for files accesed
from NFS mounted filesystems where the server returned a different
fsid or filehandle.

increase all local buffers to 1K as a workaround, so that the longer
line can be parsed and that there is no chance to overflow them
triggering "stack smashing" issues that would result in a terminated
gmond if any of the parts is longer than expected.

Signed-off-by: Jason A. Smith <smithj4@bnl.gov>