Don't mess with pidfile if it has our pid
If you HUP gmond and you ran it with -p it will reexec itself, keeping its
pid, read the pidfile, find a process with its pid running and exit (so as
to prevent two gmonds running at once).  This isn't correct as it should
just keep running and not do anything with its pidfile.