Fix bug for carbon timeout.
While trying to modify carbon_timeout using gmetad.conf
noticed timeout wasn't affecting actual socket timeout.
this was due to a bug in rrd_helpers.c while checking the
configuration value of the carbon_timeout variable in
gmetad.conf
and relying on the default 500 ms value.

Credit for the discovery of the bug goes to Leonid
Mirsky <leonidlm@gmail.com>