Fixed a logic error with my recent gzip patch
The gzip patch I submitted recently assumes that apr_socket_data_get
returns with an error code if there is no data in the socket;
however, this is not always the case. This patch explicitly checks
for the data we want instead of relying on the error code to
indicate its presence.