Add option to prevent gmetad from generating summaries for sFlow VM metrics
This patch adds a new "unsummarized_sflow_vm_metrics" config option to gmetad
to prevent generating summaries for metrics that appears to be sFlow VM metrics.

A metric is determined to be an sFlow VM metric when the metric name contains
character '.' followed by character 'v'.

Ex: xenvm1.vcpu_num is an sFlow VM metric, vnode_cpu_num is not.