Add a "summarized_metrics" option.
This feature works notionally the same as unsummarized_metrics, except
it serves as a list of metrics to summarize explicitly (instead of
ignoring metrics to leave out). This can be useful when aggregating
large numbers of metrics where only a very small number of metrics are
actually useful when summarized.

Note that while unsummarized_metrics works for filtering cluster / host
/ metrics, summarized_metrics only changes the behavior of summaries.