[Add] all the metrics in libmetrics.h are now called in server.c
Note:

gexec_func, heartbeat_fun and location_func are not implemented in libmetrics (library linking error) and are commented out for that reason.

If the linking to these functions is done, then, by simply uncommenting them and adding the type to heartbeat and location in the structure (line 108 & 110) and in the loop (line 732 & 734) they would be called properly.