hash: replace ck_rwlock with apr_thread_rwlock
When I initially introduced this code, I was a bit oblivious towards the
use case of the hash table. The overhead of the critical sections is
nontrivial due to the design, and possible number of threads contending on
the locks is relatively high, especially when aggregating from many data
sources.

CK will come back at some point, but for now, APR locks are good enough.
Most modern pthreads implementations do optimistic spinning before sending
the thread to sleep in the kernel, so this should fix issues with high CPU
usage that have been reported over the past few months on the list.