build cpu_stat along with the other python modules
At some point the default set of `pyconf` files and which modules were
actually being built got out of sync.  This meant the cpu python files
were not present but were referenced in a `pyconf` file, resulting in
`Unable to find the metric information for foo` errors when gmond was
started. (And obviously there were no metrics either.)

fix #180