do not change the timestamp of a node upon receipt of spoofed data
centrally collected out-of-band data (eg. from BMCs or a blade
chassis) which is then spoofed into the gmond network does not mean
that the node is actually up. unfortunately ganglia's notion of a node
being "up" is based upon the time when data was last received by the
gmond network - the "REPORTED" field in gmond's xml. fix this problem
by not updating the node's timestamp in gmond if the data source is
spoofed data.

users that wish to refresh a node's timestamp with spoofed data
(eg. if a node consists of only spoofed data) can use the -H option
to gmetric to explicitly send a heartbeat message and keep the node
in the "up" state.