gmond: Fix empty pollset behavior
Previously this would attempt to wait on an empty pollset if the user
for some reason set `deaf=no` without configuring a UDP receive channel.
This would lead (on Linux) to a tight loop calling `epoll_wait` which
immediately failed with `EINVAL` as we provided no fds to wait on. This
resulted in high CPU usage.