downgrade udp recv buffer length check
The 2x SO_RCVBUF behavior is specific to Linux.  When requesting the
size of a buffer an OS might reasonably say it's exactly the size you
just asked for.  This is the behavior on Illumos and FreeBSD.

To avoid a thicket of environment specific conditionals while
accommodating Linux, downgrade to a log warning instead of
EXIT_FAILURE.

NOTE: Removed the now misleading APR 0.9.4 warning.  That version was
released in 2003.